refactor: move `firestore-collection-paths.ts` into `src/` and update its import path.
